+++ pkg/analyzer/lib/src/dart/ast/to_source_visitor.dart
@@ -1019,6 +1019,7 @@ class ToSourceVisitor implements AstVisitor<void> {
     _visitNode(node.type);
   }
 
+  @Deprecated('Override visitNamedType instead')
   @override
   void visitTypeName(TypeName node) {
     throw StateError('Should not be invoked');